requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
urllib3>=2.2.0
